Use mass-weighted normal modes
Former-commit-id: 3b6443e12e3ed80eebc5d665d54167f6c08c53c3

+++ pytrans/analysis/analysis.py
@@ -82,7 +82,7 @@ def analyse_potential(trap: AbstractTrapModel, voltages: NDArray, ions: Union[Io
                       plot=True, plot_roi=None, trap_axis='x', axes=None, title='',
                       minimize_options=dict(), verbose=True):
 
-    r0 = np.asarray(r0)
+    r0 = np.asarray(r0, dtype=float)
     if isinstance(ions, Ion):
         ion1 = ions
         r_cm = r0
